+++ sonar-project.properties
@@ -86,7 +86,6 @@ sonar.coverage.exclusions=\
   **/*fixtures.ts, \
   **/*.stories.tsx, \
   src/context/Search/suggestions/oakCurriculumData.ts, \
-  src/utils/curriculum/filteringApp.ts, \
   src/node-lib/curriculum-api-2023/sdk.ts
 
 # code duplication exclusions
@@ -122,7 +121,6 @@ sonar.cpd.exclusions=\
   **/*.stories.tsx, \
   src/components/CurriculumComponents/CurricTimetablingYearCard/index.tsx, \
   src/app/(core)/teachers/programmes/[slug]/**, \
-  src/utils/curriculum/filteringApp.ts, \
   src/node-lib/curriculum-api-2023/queries/teachersUnitOverview/helpers.ts, \
   src/node-lib/curriculum-api-2023/queries/teachersUnitOverview/teachersUnitOverview.schema.ts, \
   src/node-lib/curriculum-api-2023/queries/teachersLessonOverview/teachersLessonOverview.query.ts
