+++ src/app/(core)/teachers/programmes/[slug]/[tab]/Components/ProgrammeHeader/buildProgrammeHeading.ts
@@ -7,7 +7,7 @@ import {
   childSubjectForFilter,
   subjectCategoryForFilter,
   shouldDisplayFilter,
-} from "@/utils/curriculum/filteringApp";
+} from "@/utils/curriculum/filtering";
 import { CurriculumUnitsFormattedData } from "@/pages-helpers/curriculum/docx/tab-helpers";
 
 type SubjectTitleSelection = {
