+++ src/utils/curriculum/filtersUrl.ts
@@ -304,41 +304,6 @@ export function shouldDisplayFilter(
   }
 }
 
-export function diffFilters(
-  defaultFilterFilter: CurriculumFilters,
-  filters: CurriculumFilters,
-) {
-  const out: CurriculumFilters = {
-    childSubjects: [],
-    subjectCategories: [],
-    tiers: [],
-    years: [],
-    threads: [],
-    pathways: [],
-    keystages: [],
-  };
-  for (const keyRaw of Object.keys(defaultFilterFilter)) {
-    const key = keyRaw as keyof CurriculumFilters;
-    if (!isEqual(defaultFilterFilter[key], filters[key])) {
-      out[key] = filters[key];
-    }
-  }
-  return out;
-}
-
-export function getNumberOfFiltersApplied(
-  defaultFilterFilter: CurriculumFilters,
-  filters: CurriculumFilters,
-) {
-  const diff = diffFilters(defaultFilterFilter, filters);
-  return Object.values(diff).reduce((acc, curr) => {
-    if (curr.length) {
-      return acc + 1;
-    }
-    return acc;
-  }, 0);
-}
-
 export function subjectCategoryForFilter(
   data: CurriculumUnitsFormattedData,
   filter: CurriculumFilters,
@@ -364,162 +329,6 @@ export function childSubjectForFilter(
       return yearDataItem.childSubjects;
     })
     .find((childSubject) => childSubject.subject_slug === slug);
-}
-
-export function tierForFilter(
-  data: CurriculumUnitsFormattedData,
-  filter: CurriculumFilters,
-) {
-  const slug = filter.tiers[0];
-  if (!slug) return;
-  return Object.entries(data.yearData)
-    .flatMap(([, yearDataItem]) => {
-      return yearDataItem.tiers;
-    })
-    .find((tier) => tier.tier_slug === slug);
-}
-
-export function threadForFilter(
-  data: CurriculumUnitsFormattedData,
-  filter: CurriculumFilters,
-) {
-  const slug = filter.threads[0];
-  if (!slug) return;
-  return data.threadOptions.find((thread) => thread.slug === slug);
-}
-
-export function buildTextDescribingFilter(
-  data: CurriculumUnitsFormattedData,
-  filters: CurriculumFilters,
-) {
-  const keystageSuffix = keystageSuffixForFilter(
-    data,
-    filters,
-    "subjectCategories",
-  );
-  const subjectCategoryTitle = subjectCategoryForFilter(data, filters)?.title;
-  const keystageSuffixWithSpace = keystageSuffix ? ` ${keystageSuffix}` : "";
-  const isAllCategories = filters.subjectCategories[0] === "all";
-  const subjectCategoryText = isAllCategories
-    ? "All categories"
-    : `${subjectCategoryTitle}${keystageSuffixWithSpace}`;
-  const subjectCategoryField =
-    filters.subjectCategories.length > 0 ? subjectCategoryText : undefined;
-
-  const childSubjectField = childSubjectForFilter(data, filters)?.subject;
-  const childSubjectSuffix = keystageSuffixForFilter(
-    data,
-    filters,
-    "childSubjects",
-  );
-  const childSubjectWithSuffix =
-    childSubjectField && childSubjectSuffix
-      ? `${childSubjectField} ${childSubjectSuffix}`
-      : childSubjectField;
-
-  const tierField = tierForFilter(data, filters)?.tier;
-  const tierSuffix = keystageSuffixForFilter(data, filters, "tiers");
-  const tierWithSuffix =
-    tierField && tierSuffix ? `${tierField} ${tierSuffix}` : tierField;
-
-  const fields = [
-    subjectCategoryField,
-    childSubjectWithSuffix,
-    tierWithSuffix,
-    threadForFilter(data, filters)?.title,
-  ].filter(Boolean);
-
-  return fields;
-}
-
-function getKeystageFromYear(year: number): string {
-  if (year >= 1 && year <= 2) {
-    return "(KS1)";
-  }
-  if (year >= 3 && year <= 6) {
-    return "(KS2)";
-  }
-  if (year >= 7 && year <= 9) {
-    return "(KS3)";
-  }
-  if (year >= 10 && year <= 11) {
-    return "(KS4)";
-  }
-  return "";
-}
-
-function hasFilterMatch(
-  yearData: YearData[number],
-  filterType: "subjectCategories" | "childSubjects" | "tiers",
-  filterId: string,
-): boolean {
-  if (filterType === "subjectCategories") {
-    return yearData.subjectCategories?.some((sc) => sc.slug === filterId);
-  }
-  if (filterType === "childSubjects") {
-    return yearData.childSubjects?.some((cs) => cs.subject_slug === filterId);
-  }
-  return yearData.tiers?.some((t) => t.tier_slug === filterId);
-}
-
-export function keystageSuffixForFilter(
-  data: CurriculumUnitsFormattedData,
-  filter: CurriculumFilters,
-  filterType:
-    | "subjectCategories"
-    | "childSubjects"
-    | "tiers" = "subjectCategories",
-): string | undefined {
-  if (!filter[filterType].length) {
-    return undefined;
-  }
-
-  const filterId = filter[filterType][0];
-  if (!filterId || filterId === "-1") {
-    return undefined;
-  }
-
-  const yearsWithChildSubjects: string[] = [];
-  const yearsWithoutChildSubjects: string[] = [];
-
-  Object.entries(data.yearData).forEach(([year, yearData]) => {
-    if (hasFilterMatch(yearData, filterType, filterId)) {
-      if (yearData.childSubjects?.length > 0) {
-        yearsWithChildSubjects.push(year);
-      } else {
-        yearsWithoutChildSubjects.push(year);
-      }
-    }
-  });
-
-  // Prioritise years without child subjects if available
-  const relevantYears =
-    yearsWithoutChildSubjects.length > 0
-      ? yearsWithoutChildSubjects
-      : yearsWithChildSubjects;
-
-  if (relevantYears.length === 0) {
-    return undefined;
-  }
-
-  const sortedYears = [...relevantYears].sort((a, b) => Number(a) - Number(b));
-  const firstYear = Number(sortedYears[0]);
-
-  // Check for combined key stages
-  const hasCombinedKS = sortedYears.some(
-    (year) =>
-      (Number(year) <= 2 &&
-        sortedYears.some((y) => Number(y) >= 3 && Number(y) <= 6)) ||
-      (Number(year) >= 7 &&
-        Number(year) <= 9 &&
-        sortedYears.some((y) => Number(y) >= 10 && Number(y) <= 11)),
-  );
-
-  if (hasCombinedKS) {
-    return "";
-  }
-
-  return getKeystageFromYear(firstYear);
 }
 
 export function getNumberOfSelectedUnits(
